move some tmp functions to tuple to improve compiletime
