+++ include/cib/detail/meta.hpp
@@ -22,70 +22,7 @@ namespace cib::detail {
             std::remove_cv_t<std::remove_reference_t<Rhs>>
         >;
 
-    /**
-     * Used by fold_right to leverage c++17 fold expressions with arbitrary
-     * callables.
-     *
-     * @tparam ValueType
-     *      The type of the element from the value pack.
-     *
-     * @tparam CallableType
-     *      A callable that takes two arguments, current element to be
-     *      processed and the fold state.
-     */
-    template<
-        typename ValueType,
-        typename CallableType>
-    struct fold_helper {
-        ValueType const & element_;
-        CallableType const & operation_;
 
-        CIB_CONSTEXPR fold_helper(
-            ValueType const & element,
-            CallableType const & operation
-        )
-            : element_{element}
-            , operation_{operation}
-        {}
-
-        template<typename StateType>
-        [[nodiscard]] CIB_CONSTEXPR inline auto operator+(
-            StateType const & state
-        ) const {
-            return operation_(element_, state);
-        }
-    };
-
-    /**
-     * fold_right a tuple of elements.
-     *
-     * Fold operations are sometimes called accumulate or reduce in other
-     * languages or libraries.
-     *
-     * https://en.wikipedia.org/wiki/Fold_%28higher-order_function%29
-     *
-     * @param operation
-     *      A callable that takes the current element being processed
-     *      and the current state, and returns the state to be used
-     *      to process the next element. Called for each element in
-     *      the tuple.
-     *
-     * @return
-     *      The final state of all of the operations.
-     */
-    template<
-        typename... ElementTypes,
-        typename InitType,
-        typename CallableType>
-    [[nodiscard]] CIB_CONSTEXPR inline static auto fold_right(
-        tuple_impl<ElementTypes...> const & elements,
-        InitType const & initial_state,
-        CallableType const & operation
-    ) {
-        return apply([&](auto const & ... element_pack){
-            return (fold_helper{element_pack, operation} + ... + initial_state);
-        }, elements);
-    }
 
     /**
      * Perform an operation on each element of an integral sequence.
@@ -121,24 +58,6 @@ namespace cib::detail {
     ) {
         CIB_CONSTEXPR auto seq = std::make_integer_sequence<IntegralType, NumElements>{};
         for_each(seq, operation);
-    }
-
-    /**
-     * Perform an operation on each element of a tuple.
-     *
-     * @param operation
-     *      The operation to perform. Must be a callable that accepts a single parameter.
-     */
-    template<
-        typename... ElementTypes,
-        typename CallableType>
-    CIB_CONSTEXPR inline void for_each(
-        tuple_impl<ElementTypes...> const & elements,
-        CallableType const & operation
-    ) {
-        apply([&](auto const & ... element_pack){
-            (operation(element_pack) , ...);
-        }, elements);
     }
 }
