+++ include/cib/nexus.hpp
@@ -35,7 +35,7 @@ namespace cib {
         #undef CIB_BUILD_SERVICE
 
         static void init() {
-            detail::for_each(initialized_builders_v<Config>, [](auto b){
+            initialized_builders_v<Config>.for_each([](auto b){
                 using service_tag = typename decltype(b)::Service;
                 using clean_service_tag = std::remove_cv_t<std::remove_reference_t<service_tag>>;
